0.106.1: P2 §5 — overflow guards for OpenSSL int+int size sums
P2 "Надежность и оптимизация" §5 (partial).

Audit found 4 sites in dlms-security where OpenSSL EVP_*Update/Final
returns two ints that were summed as `int + int` and then cast to
`std::size_t`:
  - lib/dlms-security/src/security/hls_high_authenticator.cpp:54
  - lib/dlms-security/src/security/suite0_aes_gcm.cpp:148 (Seal)
  - lib/dlms-security/src/security/suite0_aes_gcm.cpp:225 (Open)
  - lib/dlms-security/src/security/suite0_key_wrap.cpp:95

Formally this is signed-integer addition that UBSan would flag as
`signed-integer-overflow` at INT_MAX-adjacent inputs (not reachable
on real DLMS APDUs, but UBSan-cleanliness matters for the new
linux-sanitizers CI job).

Fix: cast each int to std::size_t before summing. OpenSSL guarantees
both are >= 0, so semantics are unchanged.

Other size arithmetic in the framework already has guards:
  - ber.cpp checks `result > size_t::max() >> 8` before each shift
  - llc_codec::HasRepresentableLpduSize
  - apdu_c_api `payload_size == size_t::max()`
  - in_memory_invocation_counter_store `nextLocal_ == uint32_t::max()`
  - suite0_aes_gcm::SizeFitsInt limits OpenSSL input to INT_MAX

Reserve sites in association/ciphered/hls_gmac/xdlms_client sum small
constants with already-validated wire sizes — overflow would require
16 EB of memory.

Patch bump: internal fixes, no API/ABI/behavior change.
Local MinGW64 ctest: 976/976 green.

docs/production_readiness_roadmap.md: P2 §5 marked DONE (partial).

+++ lib/dlms-security/src/security/hls_high_authenticator.cpp
@@ -51,7 +51,11 @@ SecurityStatus Aes128EcbEncrypt(
         &finalSize) != 1) {
     status = SecurityStatus::CipherFailed;
   } else {
-    data.resize(static_cast<std::size_t>(outputSize + finalSize));
+    // Sum as size_t to avoid signed-int overflow UB on the int+int
+    // intermediate. EVP_EncryptUpdate/Final guarantee both are >= 0.
+    data.resize(
+      static_cast<std::size_t>(outputSize) +
+      static_cast<std::size_t>(finalSize));
   }
 
   EVP_CIPHER_CTX_free(context);
